+++ Vikas_sahu/types/habits.ts
@@ -22,4 +22,3 @@ export interface HabitData {
   }
 }
 
-
